Refactor codebase folder structure
