+++ packages/cli/package.json
@@ -68,7 +68,6 @@
     "ignore": "^4.0.6",
     "inquirer": "^7.0.0",
     "internal-ip": "^5.0.0",
-    "is-ipfs": "^0.6.3",
     "listr": "^0.14.3",
     "listr-input": "0.2.0",
     "listr-silent-renderer": "^1.1.1",
@@ -116,14 +115,11 @@
     "eslint-plugin-prettier": "^3.0.0",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
-    "ganache-cli": "^6.7.0",
-    "ganache-core": "^2.9.0",
     "nyc": "^14.1.1",
     "prettier": "^1.15.3",
     "proxyquire": "^2.1.0",
     "sinon": "^7.3.1",
-    "strip-ansi": "^6.0.0",
-    "truffle": "^5.1.6"
+    "strip-ansi": "^6.0.0"
   },
   "engines": {
     "node": ">=10.0.0"
