+++ packages/toolkit/config/environments.default.json
@@ -5,35 +5,17 @@
   },
   "rinkeby": {
     "gasPrice": "15000000001",
-    "apm": {
-      "ipfs": {
-        "gateway": "https://ipfs.eth.aragon.network/ipfs"
-      }
-    },
     "registry": "0x98df287b6c145399aaa709692c8d308357bc085d",
-    "wsRPC": "wss://rinkeby.eth.aragon.network/ws",
     "network": "rinkeby"
   },
   "staging": {
     "gasPrice": "15000000001",
-    "apm": {
-      "ipfs": {
-        "gateway": "https://ipfs.eth.aragon.network/ipfs"
-      }
-    },
     "registry": "0xfe03625ea880a8cba336f9b5ad6e15b0a3b5a939",
-    "wsRPC": "wss://rinkeby.eth.aragon.network/ws",
     "network": "rinkeby"
   },
   "mainnet": {
     "gasPrice": "1500000000",
-    "apm": {
-      "ipfs": {
-        "gateway": "https://ipfs.eth.aragon.network/ipfs"
-      }
-    },
     "registry": "0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e",
-    "wsRPC": "wss://mainnet.eth.aragon.network/ws",
     "network": "mainnet"
   }
 }
