+++ packages/toolkit/package.json
@@ -8,6 +8,7 @@
     "url": "https://github.com/aragon/aragon-cli/tree/master/packages/toolkit"
   },
   "scripts": {
+    "generate": "typechain --target web3-v1 --outDir typechain './node_modules/@aragon/abis/os/artifacts/*.json'",
     "build": "babel src --out-dir dist --source-maps --extensions '.js,.ts'",
     "build:watch": "npm run build -- --watch",
     "build:docs": "typedoc --out docs --allowJs --readme none --theme minimal --mode file",
@@ -65,7 +66,7 @@
     "web3-utils": "^1.2.4"
   },
   "devDependencies": {
-    "@aragon/test-helpers": "^2.1.0",
+    "@aragon/contract-helpers-test": "^0.0.1",
     "@babel/cli": "^7.7.4",
     "@babel/core": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
@@ -76,6 +77,7 @@
     "ava": "^2.4.0",
     "babel-plugin-istanbul": "^5.2.0",
     "bn.js": "^5.0.0",
+    "bn.js-typings": "^1.0.1",
     "eslint": "^6.7.2",
     "eslint-config-prettier": "^6.7.0",
     "eslint-config-standard": "^14.1.0",
@@ -86,10 +88,12 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
     "ethereumjs-abi": "^0.6.8",
+    "ethers": "^4.0.44",
     "nyc": "^14.1.1",
     "prettier": "^1.19.1",
     "proxyquire": "^2.1.3",
     "sinon": "^7.5.0",
+    "typechain": "^1.0.5",
     "typedoc": "^0.15.4",
     "typescript": "^3.7.3"
   },
