--- packages/toolkit/src/acl/index.js
@@ -1,3 +1,0 @@
-export * from './params'
-export * from './view'
-export * from './viewFormatter'
