+++ packages/toolkit/src/apm/index.js
@@ -1,4 +1,3 @@
-export { default as useApm } from './useApm'
+export { default as getApm } from './apm'
 export { default as getApmRepo } from './getApmRepo'
 export { default as getApmRegistryPackages } from './getApmRegistryPackages'
-export { default as grantNewVersionsPermission } from './grantNewVersionsPermission'
