+++ packages/toolkit/src/dao/kernel/index.js
@@ -1,0 +1,3 @@
+export * from './kernel'
+export * from './apps'
+export { default as getAppKernel } from './getAppKernel'
