+++ packages/toolkit/src/helpers/index.js
@@ -1,7 +1,8 @@
+export * from './ipfs'
+export * from './node'
 export * from './wrapper'
 export * from './constants'
 export * from './solidity-extractor'
-export * from './generateArtifact'
 export * from './useEnvironment'
 export * from './web3-fallback'
 export * from './loadConfigFiles'
@@ -9,4 +10,3 @@ export * from './truffle-config'
 export * from './resolveAddressOrEnsDomain'
 export { default as defaultAPMName } from './default-apm'
 export { default as extractContractInfoToFile } from './extractContractInfoToFile'
-export { default as encodeInitPayload } from './encodeInitPayload'
