+++ packages/toolkit/src/index.js
@@ -1,9 +1,4 @@
-export * from './acl'
 export * from './apm'
 export * from './dao'
 export * from './helpers'
-export * from './ipfs'
-export * from './kernel'
-export * from './node'
-export * from './token/token'
-export * from './util'
+export * from './utils'
