--- packages/toolkit/src/kernel/index.js
@@ -1,2 +1,0 @@
-export { default as getAppKernel } from './getAppKernel'
-export * from './kernel'
