+++ packages/toolkit/src/util.js
@@ -92,6 +92,7 @@ export const getRecommendedGasLimit = async (
   const bufferedGasLimit = Math.round(estimatedGas * gasFuzzFactor)
 
   if (bufferedGasLimit < upperGasLimit) return bufferedGasLimit
+  console.log('>>gasLimit:', upperGasLimit)
   return upperGasLimit
 }
