+++ packages/toolkit/src/utils/encodeInitPayload.js
