+++ packages/toolkit/src/utils/token.js
