--- packages/toolkit/test/test-helpers/isValidTxHash.js
@@ -1,1 +1,0 @@
-export const isValidTxHash = txHash => /^0x([A-Fa-f0-9]{64})$/.test(txHash)
